poco a poco cambios internos

+++ Practica1/src/clases/Ecosistema.java
@@ -23,12 +23,12 @@ public class Ecosistema {
 		
 	}
 	
-	public static double distancia(ElementoEcosistema ee, ElementoEcosistema ee1) {
+	public static int distancia(ElementoEcosistema ee, ElementoEcosistema ee1) {
 		
 		double x = ee.getPunto().getX() - ee1.getPunto().getX();
 		double y = ee.getPunto().getY() - ee1.getPunto().getY();
 		double sumaCuadrados = x*x  + y*y;
-		double dis = Math.sqrt(sumaCuadrados);
+		int dis = (int)Math.sqrt(sumaCuadrados);
 		return dis;
 		
 	}
